+++ src/components/index.js
@@ -4,7 +4,7 @@ export { default as Home } from "./Home";
 export { default as About } from "./About";
 export { default as Contact } from "./Contact";
 export { default as Blog } from "./blog/Blog";
-export { default as CreatePost } from "./CreatePost";
+export { default as CreatePost } from "./CreateTicket";
 export { default as Login } from "./Login"
 export { default as Posts } from "./blog/Posts";
 export { default as Post } from "./blog/Post";
